requirements.txt
bkoauth==0.0.22                      # ieod requirment
bkstorages==0.1.3                    # ieod requirment
blueking-component-ieod==0.0.90      # internal
blueapps==3.2.7                      # self-developed

# need move to 01_requirements.txt on pack
# move to "02_requirements_local.txt" if end with "self-developed" tag
idna==2.8
django-mysql==3.3.0
pyppeteer==0.2.5
Django==1.11.23
Mako==1.0.6
MarkupSafe==1.1.1
PyJWT==1.5.2
arrow==0.6.0
six==1.14.0
celery==3.1.25
django-celery==3.2.1
django-filter==1.0.4
django-redis==4.8.0
djangorestframework==3.6.2
django-rest-framework-condition==0.1.1
redis==2.10.5
requests==2.22.0
sqlparse==0.2.3
ujson==1.35
schematics==2.0.1
pygments==2.2.0
pycrypto==2.6.1
Pillow==6.2.2
PyYAML==3.12
PyMySQL==0.6.7
coreapi==2.3.3
elasticsearch==5.5.3
elasticsearch-curator==5.5.4
gunicorn==19.6.0
kafka-python==1.4.6
babel==2.5.1
python-consul==1.1.0
ipython==7.13.0
netifaces==0.10.6
psutil==5.2.2
raven==6.9.0
influxdb==5.2.0
selenium==3.141.0
kazoo==2.5.0
whitenoise==3.3.0
python-json-logger==0.1.7
ddtrace==0.14.1
prometheus-client==0.6.0
graph-exporter==2.1.2                 # self-developed
dnspython==1.16.0
furl==2.0.0
sentry-sdk==0.9.0
jinja2==2.10.1
django-jinja==2.4.1
django-versionlog==1.3.0            # self-developed
humanize==0.5.1
gevent==1.1.1
celery-redbeat==0.13.0
mysqlclient==1.4.4
supervisor==4.0.2    # backend
jedi==0.17.2

# jwt
pycrypto==2.6.1
PyJWT==1.5.2
cryptography==2.7
python-magic==0.4.15

# iam
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
curlify==2.2.1

# bk-dataview
dataclasses==0.6
